Move fixtures out of test_rgba
