+++ tests/test_rgba.py
@@ -4,30 +4,8 @@ import xarray as xr
 
 from odc.geo._interop import is_dask_collection
 from odc.geo._rgba import is_rgb
-from odc.geo.data import ocean_geom
-from odc.geo.geobox import GeoBox
 from odc.geo.testutils import daskify
-from odc.geo.xr import ODCExtensionDa, ODCExtensionDs, rasterize
-
-
-@pytest.fixture()
-def ocean_raster() -> xr.DataArray:
-    gbox = GeoBox.from_bbox(bbox=(-180, -90, 180, 90), shape=(128, 256))
-    return rasterize(ocean_geom(), gbox)
-
-
-@pytest.fixture()
-def ocean_raster_ds(ocean_raster: xr.DataArray) -> xr.Dataset:
-    xx = ocean_raster.astype("int16") * 3_000
-    xx.attrs["nodata"] = -1
-
-    return xr.Dataset(
-        dict(
-            red=xx,
-            green=xx,
-            blue=xx,
-        )
-    )
+from odc.geo.xr import ODCExtensionDa, ODCExtensionDs
 
 
 def test_colorize(ocean_raster: xr.DataArray):
